pyproject.toml
[project]
name = "drf-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Rakhimzhanov Aman",email = "amanrahim2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "django-filter (>=25.2,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
